Editorial: Remove any references to "parsable MIME type"
The term "parsable MIME type" used to be part of the MIME Sniffing
standard, but it was removed in whatwg/mimesniff#36. This change
replaces its uses with equivalent phrasing that references the "parse a
MIME type" algorithm. It also replaces mentions of "ASCII-encoded
strings" with the Infra standard's definition of "ASCII string".

Closes #170.